2.3. Введение в Maven и Spring Boot (Сделал ДЗ) внес исправления, добавил челочисленное деление

+++ src/main/java/comexamplecalculator/service/CounterService.java
@@ -7,5 +7,5 @@ public interface CounterService {
 
     int multiplication(int a, int b);
 
-    int division(int a, int b);
+    double division(int a, int b);
 }
